Update tests to use unstable_mockModule
